Add ability to configure oslo_middleware
Add a class oslo_middleware, which takes in a param hash to be passed to
the oslo::middleware class. This sets up the [oslo_middleware] config
section in magnum config.

This method makes it more generic; if oslo::middleware has a new param
we do not have to change it in all different projects. This is modelled
after Change-Id I695834ac03a52d8569e50db600676a89e165491d

Change-Id: I8556c9184a77473e319b6b99148b6d3793dc1988