Add ability to configure oslo_middleware
Add a param name oslo_middleware_params, which takes in a hash to be
passed to the oslo::middleware class to set up the [oslo_middleware]
config section.

This method makes it more generic; if oslo::middleware has a new param
we do not have to change it in all different projects. This is modelled
after Change-Id I695834ac03a52d8569e50db600676a89e165491d

Change-Id: I9f7a3534e6c5d5e67145652ad21ccedcc1da7c38